LRCI-3385 Change item id to a long instead of an int

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/JenkinsResultsParserUtil.java
@@ -218,7 +218,7 @@ public class JenkinsResultsParserUtil {
 	}
 
 	public static void cancelQueuedItem(
-		int itemID, JenkinsMaster jenkinsMaster) {
+		long itemID, JenkinsMaster jenkinsMaster) {
 
 		StringBuilder sb = new StringBuilder();
 
@@ -3556,7 +3556,7 @@ public class JenkinsResultsParserUtil {
 
 			toString(sb.toString());
 
-			cancelQueuedItem(itemJSONObject.getInt("id"), jenkinsMaster);
+			cancelQueuedItem(itemJSONObject.getLong("id"), jenkinsMaster);
 		}
 	}
